requirements.txt
# requirements.txt
langgraph
langchain-groq
langchain-core
python-telegram-bot
asyncpg
python-dotenv==1.0.0